+++ NumFlat/Clustering/DIagonalGaussianMixtureModel.cs
@@ -89,10 +89,6 @@ namespace NumFlat.Clustering
                 }
                 this.components = curr.Model.components;
             }
-            catch (FittingFailureException)
-            {
-                throw;
-            }
             catch (Exception e)
             {
                 throw new FittingFailureException("Failed to fit the model.", e);
@@ -192,10 +188,6 @@ namespace NumFlat.Clustering
                 var nextComponents = nextWeights.Zip(nextGaussians, (weight, gaussian) => new Component(weight, gaussian)).ToArray();
                 return (new DiagonalGaussianMixtureModel(nextComponents), likelihood);
             }
-            catch (FittingFailureException)
-            {
-                throw;
-            }
             catch (Exception e)
             {
                 throw new FittingFailureException("Failed to fit the model.", e);
@@ -219,6 +211,29 @@ namespace NumFlat.Clustering
         public double Pdf(in Vec<double> x)
         {
             return Math.Exp(LogPdf(x));
+        }
+
+        /// <inheritdoc/>
+        public void Generate(Random random, in Vec<double> destination)
+        {
+            ThrowHelper.ThrowIfNull(random, nameof(random));
+            ThrowHelper.ThrowIfEmpty(destination, nameof(destination));
+            MultivariateDistribution.ThrowIfInvalidSize(this, destination, nameof(destination));
+
+            var target = random.NextDouble();
+            var position = 0.0;
+            var choice = components[0];
+            foreach (var component in components)
+            {
+                position += component.Weight;
+                if (position > target)
+                {
+                    choice = component;
+                    break;
+                }
+            }
+
+            choice.Gaussian.Generate(random, destination);
         }
 
         /// <summary>