+++ NumFlat/Distributions/Gaussian.cs
@@ -71,11 +71,14 @@ namespace NumFlat.Distributions
             try
             {
                 (mean, covariance) = xs.MeanAndCovariance(0);
-                Special.IncreaseDiagonalElementsInplace(covariance, regularization);
             }
             catch (Exception e)
             {
                 throw new FittingFailureException("Failed to compute the covariance matrix.", e);
+            }
+            foreach (ref var value in covariance.EnumerateDiagonalElements())
+            {
+                value += regularization;
             }
 
             var (cholesky, logNormalizationTerm) = GetCholeskyAndLogNormalizationTerm(mean, covariance);
@@ -117,11 +120,14 @@ namespace NumFlat.Distributions
             try
             {
                 (mean, covariance) = xs.MeanAndCovariance(weights, 0);
-                Special.IncreaseDiagonalElementsInplace(covariance, regularization);
             }
             catch (Exception e)
             {
                 throw new FittingFailureException("Failed to compute the covariance matrix.", e);
+            }
+            foreach (ref var value in covariance.EnumerateDiagonalElements())
+            {
+                value += regularization;
             }
 
             var (cholesky, logNormalizationTerm) = GetCholeskyAndLogNormalizationTerm(mean, covariance);
@@ -168,6 +174,33 @@ namespace NumFlat.Distributions
         public double Pdf(in Vec<double> x)
         {
             return Math.Exp(LogPdf(x));
+        }
+
+        /// <inheritdoc/>
+        public void Generate(Random random, in Vec<double> destination)
+        {
+            ThrowHelper.ThrowIfNull(random, nameof(random));
+            ThrowHelper.ThrowIfEmpty(destination, nameof(destination));
+            MultivariateDistribution.ThrowIfInvalidSize(this, destination, nameof(destination));
+
+            using var utmp = new TemporalVector<double>(mean.Count);
+            ref readonly var tmp = ref utmp.Item;
+
+            foreach (ref var value in tmp)
+            {
+                value = random.NextGaussian();
+            }
+
+            var i = 0;
+            foreach (ref var value in destination)
+            {
+                var x = cholesky.L.Rows[i].Subvector(0, i + 1);
+                var y = tmp.Subvector(0, i + 1);
+                value = x * y;
+                i++;
+            }
+
+            destination.AddInplace(mean);
         }
 
         /// <summary>
@@ -245,13 +278,12 @@ namespace NumFlat.Distributions
 
         private static double LogDeterminant(in Mat<double> l)
         {
-            var fl = l.GetUnsafeFastIndexer();
-            var value = 0.0;
-            for (var i = 0; i < l.RowCount; i++)
+            var sum = 0.0;
+            foreach(var value in l.EnumerateDiagonalElements())
             {
-                value += Math.Log(fl[i, i]);
+                sum += Math.Log(value);
             }
-            return 2 * value;
+            return 2 * sum;
         }
 
         /// <summary>
@@ -263,6 +295,11 @@ namespace NumFlat.Distributions
         /// Gets the covariance matrix.
         /// </summary>
         public ref readonly Mat<double> Covariance => ref covariance;
+
+        /// <summary>
+        /// Gets the Cholesky decomposition of the covariance matrix.
+        /// </summary>
+        public CholeskyDecompositionDouble CholeskyDecomposition => cholesky;
 
         /// <inheritdoc/>
         public int Dimension => mean.Count;