+++ NumFlat/MatrixDecompositions/QrDecompositionComplex.cs
@@ -125,11 +125,10 @@ namespace NumFlat
                 throw new InvalidOperationException("Calling this method against a non-square QR decomposition is not allowed.");
             }
 
-            var fr = r.GetUnsafeFastIndexer();
             var determinant = 1.0;
-            for (var i = 0; i < r.RowCount; i++)
+            foreach(var value in r.EnumerateDiagonalElements())
             {
-                determinant *= fr[i, i].Magnitude;
+                determinant *= value.Magnitude;
             }
             return determinant;
         }
@@ -147,11 +146,10 @@ namespace NumFlat
                 throw new InvalidOperationException("Calling this method against a non-square QR decomposition is not allowed.");
             }
 
-            var fr = r.GetUnsafeFastIndexer();
             var logDeterminant = 0.0;
-            for (var i = 0; i < r.RowCount; i++)
+            foreach (var value in r.EnumerateDiagonalElements())
             {
-                logDeterminant += Math.Log(fr[i, i].Magnitude);
+                logDeterminant += Math.Log(value.Magnitude);
             }
             return logDeterminant;
         }