+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# v0.10.2
+
+* Fixed issue where resampling failed on long signal.
+
+# v0.10.1
+
+* Added `MultivariateDistribution.Generate()` method without args which uses `Random.Shared`.
+* Added resampling.
+
+# v0.10.0
+
+* Audio feature extraction supports complex spectrum.
+* Added random sampling method for distributions.
+* Some code cleanup.
+
 # v0.9.7
 
 * Added options for clustering algorithms.